# apps/api/test/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/properties/{propertyId}": {
      "get": {
        "operationId": "PropertiesController_get",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      },
      "patch": {
        "operationId": "PropertiesController_update",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/api/v1/properties/{propertyId}/integrations/connectors": {
      "get": {
        "operationId": "IntegrationInstancesController_connectors",
        "parameters": [
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations": {
      "get": {
        "operationId": "IntegrationInstancesController_list",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      },
      "post": {
        "operationId": "IntegrationInstancesController_create",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}": {
      "get": {
        "operationId": "IntegrationInstancesController_get",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      },
      "patch": {
        "operationId": "IntegrationInstancesController_update",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInstanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/enrollment-tokens": {
      "post": {
        "operationId": "IntegrationInstancesController_enrollmentToken",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/agent": {
      "get": {
        "operationId": "IntegrationInstancesController_agent",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/agent/revoke": {
      "post": {
        "operationId": "IntegrationInstancesController_revokeAgent",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeAgentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/messages": {
      "get": {
        "operationId": "IntegrationInstancesController_messages",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "RECEIVED",
                "PROCESSED",
                "PENDING_MAPPING",
                "HELD",
                "FAILED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/messages/replay-pending": {
      "post": {
        "operationId": "IntegrationInstancesController_replayPending",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integrations/{instanceId}/messages/{messageId}/replay": {
      "post": {
        "operationId": "IntegrationInstancesController_replay",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationInstances"
        ]
      }
    },
...
    "/api/v1/properties/{propertyId}/integration-exceptions": {
      "get": {
        "operationId": "IntegrationQueueController_listExceptions",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "instanceId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "RESOLVED",
                "IGNORED"
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationQueue"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/integration-exceptions/{exceptionId}/close": {
      "post": {
        "operationId": "IntegrationQueueController_closeException",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exceptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseExceptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationQueue"
        ]
      }
    },
    "/api/v1/properties/{propertyId}/external-references": {
      "get": {
        "operationId": "IntegrationQueueController_references",
        "parameters": [
          {
            "name": "propertyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entityType",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^[a-z]+\\.[a-z_]+$",
              "type": "string"
            }
          },
          {
            "name": "entityId",
            "required": true,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "IntegrationQueue"
        ]
      }
    },
...
  },
  "info": {
    "title": "Hotella Platform API",
    "description": "Hotel Intelligence Platform — staff, guest and integration APIs.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdatePropertyDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "timezone": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "defaultLocale": {
            "type": "string",
            "pattern": "^[a-z]{2}(-[A-Z]{2})?$"
          },
          "enabledLocales": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-z]{2}(-[A-Z]{2})?$"
            }
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "nullable": true
          },
          "address": {
            "type": "object",
            "additionalProperties": {}
          },
          "settings": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "version"
        ]
      },
...
      "CreateInstanceDto": {
        "type": "object",
        "properties": {
          "connectorCode": {
            "type": "string",
            "pattern": "^[A-Z][A-Z0-9_]{1,47}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "config": {
            "default": {},
            "type": "object",
            "additionalProperties": {}
          },
          "credentialRefs": {
            "default": {},
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9+.-]*:\\/\\/\\S+$"
            }
          },
          "capabilities": {
            "minItems": 1,
            "maxItems": 64,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "RESERVATION_READ",
                "GUEST_READ",
                "CHECKIN_EVENT",
                "CHECKOUT_EVENT",
                "ROOM_MOVE_EVENT",
                "PROFILE_EVENT",
                "ROOM_STATUS_READ",
                "ROOM_STATUS_WRITE",
                "OOO_READ",
                "OOO_WRITE",
                "RECONCILIATION_READ",
                "POST_CHARGE",
                "MENU_READ",
                "ITEM_AVAILABILITY",
                "ORDER_CREATE",
                "ORDER_STATUS",
                "CHECK_READ",
                "ITEM_READ",
                "STOCK_READ",
                "REQUISITION_CREATE",
                "PO_STATUS_READ",
                "CALL_RECEIVED",
                "CALL_TRANSFER",
                "CALL_HOLD",
                "CALL_END",
                "EXTENSION_DIRECTORY",
                "WIFI_SESSION_CREATE",
                "WIFI_SESSION_REVOKE",
                "DEVICE_READ"
              ]
            }
          }
        },
        "required": [
          "connectorCode",
          "name",
          "capabilities"
        ]
      },
      "UpdateInstanceDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "PAUSED",
              "DISABLED"
            ]
          },
          "config": {
            "type": "object",
            "additionalProperties": {}
          },
          "credentialRefs": {
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9+.-]*:\\/\\/\\S+$"
            }
          },
          "capabilities": {
            "minItems": 1,
            "maxItems": 64,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "RESERVATION_READ",
                "GUEST_READ",
                "CHECKIN_EVENT",
                "CHECKOUT_EVENT",
                "ROOM_MOVE_EVENT",
                "PROFILE_EVENT",
                "ROOM_STATUS_READ",
                "ROOM_STATUS_WRITE",
                "OOO_READ",
                "OOO_WRITE",
                "RECONCILIATION_READ",
                "POST_CHARGE",
                "MENU_READ",
                "ITEM_AVAILABILITY",
                "ORDER_CREATE",
                "ORDER_STATUS",
                "CHECK_READ",
                "ITEM_READ",
                "STOCK_READ",
                "REQUISITION_CREATE",
                "PO_STATUS_READ",
                "CALL_RECEIVED",
                "CALL_TRANSFER",
                "CALL_HOLD",
                "CALL_END",
                "EXTENSION_DIRECTORY",
                "WIFI_SESSION_CREATE",
                "WIFI_SESSION_REVOKE",
                "DEVICE_READ"
              ]
            }
          }
        },
        "required": [
          "version"
        ]
      },
      "RevokeAgentDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 3,
            "maxLength": 500
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "CloseExceptionDto": {
        "type": "object",
        "properties": {
          "version": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "status": {
            "type": "string",
            "enum": [
              "RESOLVED",
              "IGNORED"
            ]
          },
          "resolution": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "version",
          "status",
          "resolution"
        ]
      },
...
    }
  }
}